Moved BatchInfo from streaming.scheduler to streaming.ui
