+++ streaming/src/main/scala/org/apache/spark/streaming/ui/BatchInfo.scala
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.spark.streaming.scheduler
+package org.apache.spark.streaming.ui
 
 import org.apache.spark.streaming.Time
+import org.apache.spark.streaming.scheduler.ReceivedBlockInfo
 
 /**
  * Class having information on completed batches.
